stabilize and speed up tests
it is not guarantueed that the check method of the retentionstrategy
is called every minute. Most of the time it is only called every 2
minutes. This slows down the tests and when the timing is too tight
can lead to test failures.
When waiting for things (builds, maintenance start or end) trigger the
call to the check method manually every 10s this allows to set shorter
waiting times in the tests.